requirements.txt
# Data & API
requests==2.32.3
httpx==0.27.2
websockets==13.1
python-binance==1.0.22
pandas==2.2.3
numpy==1.26.4
pyarrow==18.0.0

# ML - Baseline
scikit-learn==1.5.2
lightgbm==4.5.0
xgboost==2.1.2

# ML - Advanced (Phase C)
torch==2.5.0
pytorch-forecasting==1.1.1
darts==0.31.0

# RL (Phase C)
stable-baselines3==2.4.0
gymnasium==1.0.0

# Dashboard
fastapi==0.115.0
uvicorn[standard]==0.32.0
jinja2==3.1.4
plotly==5.24.1

# Storage
sqlalchemy==2.0.36

# Dev
python-dotenv==1.0.1
jupyter==1.1.1
matplotlib==3.9.2
seaborn==0.13.2
